apps/api/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
google-cloud-vision==3.5.0
openai==1.12.0
cohere==4.47
firebase-admin==6.4.0
python-multipart==0.0.9 